requirements.txt
Django==5.2
django-environ==0.11.2
django-cors-headers==4.3.1
celery==5.3.4
redis==5.0.1
Pillow==10.1.0
django-extensions==3.2.3
psycopg2-binary==2.9.9
djangorestframework==3.14.0
django-rest-framework-simplejwt==5.3.0
python-decouple==3.8
requests==2.31.0
